+++ src/apps/settings/modules/resources/IconPacks.tsx
@@ -1,7 +1,7 @@
 import { invoke, SeelenCommand } from '@seelen-ui/lib';
 import { Icon } from '@shared/components/Icon';
 import { path } from '@tauri-apps/api';
-import { Button, Switch, Tooltip } from 'antd';
+import { Button, Switch } from 'antd';
 import { Reorder } from 'framer-motion';
 import { useTranslation } from 'react-i18next';
 import { useDispatch, useSelector } from 'react-redux';
@@ -42,22 +42,20 @@ export function IconPacksView() {
     <div className={cs.list}>
       <SettingsGroup>
         <SettingsOption>
-          <b>{t('general.icon_pack.label')}</b>
-          <Tooltip title={t('general.icon_pack.open_folder')}>
-            <Button
-              type="default"
-              onClick={async () => {
-                const dataDir = await path.appDataDir();
-                invoke(SeelenCommand.OpenFile, { path: await path.join(dataDir, 'icons') });
-              }}
-            >
-              <Icon iconName="PiFoldersDuotone" />
-            </Button>
-          </Tooltip>
+          <b>{t('resources.open_folder')}</b>
+          <Button
+            type="default"
+            onClick={async () => {
+              const dataDir = await path.appDataDir();
+              invoke(SeelenCommand.OpenFile, { path: await path.join(dataDir, 'icons') });
+            }}
+          >
+            <Icon iconName="PiFoldersDuotone" />
+          </Button>
         </SettingsOption>
       </SettingsGroup>
 
-      <b>Enabled</b>
+      <b>{t('general.icon_pack.selected')}</b>
       <Reorder.Group values={_active} onReorder={onReorder} className={cs.reorderGroup}>
         {enabled.map((iconPack) => (
           <Reorder.Item key={iconPack.id} value={iconPack.metadata.filename}>
@@ -77,7 +75,7 @@ export function IconPacksView() {
         ))}
       </Reorder.Group>
 
-      <b>Disabled</b>
+      <b>{t('general.icon_pack.available')}</b>
       {disabled.map((iconPack) => (
         <ResourceCard
           key={iconPack.id}