+++ src/background/state/application/themes.rs
@@ -26,6 +26,7 @@ impl FullState {
                     if theme.id.starts_with("@deprecated") {
                         continue;
                     }
+                    theme.metadata.bundled = path.starts_with(SEELEN_COMMON.bundled_themes_path());
                     theme.metadata.filename = entry.file_name().to_string_lossy().to_string();
                     self.themes.insert(theme.metadata.filename.clone(), theme);
                 }